unit test first attempt
