+++ back/backend/handlers/auth.py
@@ -1,5 +1,6 @@
 import datetime
 import os
+import sys
 
 from flask import Blueprint, url_for, redirect, request, session
 from google.cloud import ndb
@@ -8,8 +9,13 @@ from backend.lib.secrets import get_secret
 from backend.models.user import User, Roles
 from backend.models.wca.person import Person
 from backend.models.wca.rank import RankSingle, RankAverage
+from backend.test.mock_ndb_client import ndb_client
 
-client = ndb.Client()
+if "pytest" in sys.modules:
+    client = ndb_client
+else:
+    client = ndb.Client()
+
 
 def create_bp(oauth):
   bp = Blueprint('auth', __name__)
