+++ back/backend/handlers/champions_table.py
@@ -1,3 +1,5 @@
+import sys
+
 from flask import Blueprint, render_template
 from google.cloud import ndb
 
@@ -6,9 +8,14 @@ from backend.models.champion import Champion
 from backend.models.region import Region
 from backend.models.province import Province
 from backend.models.wca.event import Event
+from backend.test.mock_ndb_client import ndb_client
 
 bp = Blueprint('champions_table', __name__)
-client = ndb.Client()
+if "pytest" in sys.modules:
+    client = ndb_client
+else:
+    client = ndb.Client()
+
 
 
 #@bp.route('/async/champions_by_year/<event_id>/<championship_type>/<championship_region>')
