+++ back/backend/handlers/regional.py
@@ -1,6 +1,8 @@
 import datetime
 import os
 import logging
+import sys
+
 import requests
 
 from flask import Blueprint, render_template, abort
@@ -11,9 +13,14 @@ from backend.models.championship import Championship
 from backend.models.region import Region
 from backend.models.province import Province
 from backend.models.user import User
+from backend.test.mock_ndb_client import ndb_client
 
 bp = Blueprint('regional', __name__)
-client = ndb.Client()
+if "pytest" in sys.modules:
+    client = ndb_client
+else:
+    client = ndb.Client()
+
 
 #@bp.route('/regional')
 def regional():
