+++ back/backend/handlers/user.py
@@ -1,14 +1,20 @@
 from flask import Blueprint, jsonify, request
 from google.cloud import ndb
 import datetime
+import sys
 
 from backend.lib import permissions, auth
 from backend.models.province import Province
 from backend.models.user import User, UserLocationUpdate
 from backend.models.wca.rank import RankSingle, RankAverage
+from backend.test.mock_ndb_client import ndb_client
 
 bp = Blueprint('user', __name__)
-client = ndb.Client()
+
+if "pytest" in sys.modules:
+    client = ndb_client
+else:
+    client = ndb.Client()
 
 
 @bp.route('/user_info', methods=['GET'])
